+++ RELEASE_NOTES.md
@@ -3,10 +3,12 @@
 # MongoDB Enterprise Kubernetes Operator 1.11.1
 
 ## MongoDB Resource
-* Bug Fixes
+* Changes
   * If `spec.exposedExternally` is set to `false` after being set to `true`, the Operator will now delete the corresponding service
 
 ## MongoDBOpsManager Resource
+* Changes
+  * If `spec.externalConnectivity` is unset after being set, the Operator will now delete the corresponding service
 * Ops Manager versions 4.4.13 and 4.2.25 are now supported.
 
 ## Miscellaneous